Redesigned Search Fucntion
